requirements.txt
pandas
plotly-express
requests
fpdf
